move check to ComplexNewAnalyzer
